cleaned up the code and changed the return of vectors to shared pointers to avoid copying

+++ mu1664al-s/src/libclientserver/dbinterface.h
@@ -13,7 +13,7 @@ using std::vector;
 
 struct Article
 {
-	int id; // unique
+	size_t id; // unique
 	string title;
 	string author;
 	string text;
@@ -21,7 +21,7 @@ struct Article
 
 struct Group
 {
-	int id;		  // unique
+	size_t id;	  // unique
 	string title; // unique
 	vector<Article> articles;
 };
@@ -45,19 +45,19 @@ class DBInterface
 public:
 	virtual ~DBInterface() = default;
 
-	virtual void writeArticle(int group, const Article &article) = 0; //write a new article to a newsgroup
+	virtual void writeArticle(size_t group, const Article &article) = 0; //write a new article to a newsgroup
 
-	virtual const Article readArticle(int group, int article) = 0; // read an article from a newsgroup, using group identity and article identity
+	virtual const Article readArticle(size_t group, size_t article) = 0; // read an article from a newsgroup, using group identity and article identity
 
 	virtual const std::shared_ptr<vector<Group>> readGroups() = 0;
 
 	virtual void writeGroup(const string &name) = 0; //
 
-	virtual const std::shared_ptr<vector<Article>> readArticles(int group) = 0; // get articles from a newsgroup, using the group identity
+	virtual const std::shared_ptr<vector<Article>> readArticles(size_t group) = 0; // get articles from a newsgroup, using the group identity
 
-	virtual void deleteArticle(int group, int article) = 0; //delete an article, specified by its id in a newsgroup.
+	virtual void deleteArticle(size_t group, size_t article) = 0; //delete an article, specified by its id in a newsgroup.
 
-	virtual void deleteGroup(int group) = 0; //deletegroup, using the group identity.
+	virtual void deleteGroup(size_t group) = 0; //deletegroup, using the group identity.
 };
 
 #endif
