+++ mu1664al-s/src/libclientserver/message.h
@@ -10,7 +10,7 @@ using std::vector;
 struct Parameter
 {
     const Protocol type;
-    const int N;
+    const size_t N;
     const string str;
 };
 
@@ -31,7 +31,7 @@ public:
     Message() = default;
     Message(const string &package); // parse and validate message package
     const string encode() const;
-    Message &addNumParam(int num);
+    Message &addNumParam(size_t num);
     Message &addStrParam(const string &str);
     Message &setCommand(const Protocol &command)
     {
@@ -66,6 +66,6 @@ private:
     vector<Parameter> parameters{};
 
     int decNum(const string &str) const;
-    const string encNum(int num) const;
+    const string encNum(size_t num) const;
     string encodeParams() const;
 };
